refactor(admin): move admin page to admin folder
