--- src/app/(admin)/AdminLayout/showcase-section.tsx
@@ -1,20 +1,0 @@
-import { cn } from "@/lib/utils";
-import type { ReactNode } from "react";
-
-type PropsType = {
-  title: string;
-  children: ReactNode;
-  className?: string;
-};
-
-export function ShowcaseSection({ title, children, className }: PropsType) {
-  return (
-    <div className="rounded-[10px] bg-white shadow-1 dark:bg-gray-dark dark:shadow-card hover:shadow-lg dark:hover:shadow-xl transition-all duration-300 hover:translate-y-[-2px]">
-      <h2 className="border-b border-stroke px-4 py-4 font-medium text-dark dark:border-dark-3 dark:text-white sm:px-6 xl:px-7.5">
-        {title}
-      </h2>
-
-      <div className={cn("p-4 sm:p-6 xl:p-10", className)}>{children}</div>
-    </div>
-  );
-}
