+++ src/app/admin/components/layout/AdminLayout/index.tsx
@@ -1,0 +1,31 @@
+"use client";
+
+import { ReactNode } from "react";
+import { Header } from "./header";
+import { Sidebar } from "./sidebar";
+import { SidebarProvider } from "./sidebar/sidebar-context";
+
+interface AdminLayoutProps {
+  children: ReactNode;
+}
+
+export function AdminLayout({ children }: AdminLayoutProps) {
+  return (
+    <SidebarProvider defaultOpen={true}>
+      <div className="flex min-h-screen bg-gray-50">
+        <Sidebar />
+
+        <div className="flex flex-1 flex-col">
+          <Header />
+
+          <main className="flex-1 p-4 md:p-5 2xl:p-10">
+            {children}
+          </main>
+        </div>
+      </div>
+    </SidebarProvider>
+  );
+}
+
+export default AdminLayout;
+
