+++ src/app/admin/components/layout/AdminLayout/sidebar/data/ui-elements-list.ts
