+++ src/app/admin/dashboard/page.tsx
@@ -1,4 +1,4 @@
-'use client';
+    'use client';
 import React from 'react';
 import { 
   LayoutDashboard,
