+++ src/app/admin/layout.tsx
@@ -1,0 +1,2 @@
+
+
