+++ src/app/layout.tsx
@@ -1,0 +1,39 @@
+'use client';
+import { Geist, Geist_Mono } from "next/font/google";
+import "./globals.css";
+import { Provider } from 'react-redux';
+import { store } from '../application/redux/store';
+
+const geistSans = Geist({
+    variable: "--font-geist-sans",
+    subsets: ["latin"],
+});
+
+const geistMono = Geist_Mono({
+    variable: "--font-geist-mono",
+    subsets: ["latin"],
+});
+
+export default function RootLayout({
+                                       children,
+                                   }: Readonly<{
+    children: React.ReactNode;
+}>) {
+    return (
+        <html lang="vi" suppressHydrationWarning>
+        <head>
+            <title>🎬 Studio Booking - Thuê Studio Chuyên Nghiệp</title>
+            <meta name="description" content="Hệ thống đặt lịch thuê studio chuyên nghiệp cho chụp ảnh, quay phim, livestream và workshop." />
+            <meta name="viewport" content="width=device-width, initial-scale=1" />
+        </head>
+        <body
+            className={`${geistSans.variable} ${geistMono.variable} antialiased`}
+            suppressHydrationWarning
+        >
+        <Provider store={store}>
+            {children}
+        </Provider>
+        </body>
+        </html>
+    );
+}
